feat: implement configuration management and TUI integration
